refactor: rename drop position-related files
