+++ packages/editor/src/editor/render.element.tsx
@@ -3,7 +3,7 @@ import {useSelector} from '@xstate/react'
 import {useContext, type ReactElement} from 'react'
 import type {Element as SlateElement} from 'slate'
 import {useSlateStatic, type RenderElementProps} from 'slate-react'
-import type {DropPosition} from '../behaviors/behavior.core.block-element'
+import type {DropPosition} from '../behaviors/behavior.core.drop-position'
 import type {
   RenderBlockFunction,
   RenderChildFunction,
